requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-environ==0.4.5
django-filter==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
idna==2.8
Markdown==3.1
oauthlib==3.0.1
Pillow==6.0.0
psycopg2==2.8.2
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
